split import component to service and uniq responsability
